feature: finalização da implementação da venda dos produtos com validações.

+++ RO.DevTest.Domain/Entities/SaleItem.cs
@@ -11,7 +11,7 @@ namespace RO.DevTest.Domain.Entities
         public Product Product { get; set; } = null!;
 
         public int Quantidade { get; set; }
-        public decimal PrecoUnitario { get; set; }
+        public float PrecoUnitario { get; set; }
     }
 
 }
